Drop gcc from Depends.
gcc was added for the purpose of running the use.t autopkgtest (cf. #784846);
in the meantime it would need at least g++, libc6-dev, etc., which would
increase the installation size of (transitive) reverse dependencies even
more.

As the build environment is only needed when libalien-wxwidgets-perl is
required as a build dependency, and the necessary build tools are present in
this situation anyway, the direct runtime dependency can be dropped.

Cf. also #797650 in libwx-perl.